set jwt

+++ src/services/product.service.ts
@@ -1,6 +1,6 @@
 import { api } from "@components/api"
 import { IProduct } from "@components/interfaces/product.interface"
-
+import { authorizationHeader } from "../functions/authorization-header"
 class ProductService {
     async getAllProducts() {
         const response = await api.get("/product")
@@ -13,17 +13,17 @@ class ProductService {
     }
 
     async createProduct(product: IProduct) {
-        const response = await api.post("/product", product)
+        const response = await api.post("/product", product, authorizationHeader())
         return response.data
     }
 
     async updateProduct(product: IProduct) {
-        const response = await api.put(`/product/${product._id}`, product)
+        const response = await api.put(`/product/${product._id}`, product, authorizationHeader())
         return response.data
     }
 
     async removeProduct(id: string) {
-        const response = await api.delete(`/product/${id}`)
+        const response = await api.delete(`/product/${id}`, authorizationHeader())
         return response.data
 
     }
